refactor: Move callbackDispatcher function to it's own separate file
